Update welcome.yml
#3
The step you've added uses GitHub CLI (gh) to add a comment when a pull request is opened. To allow GitHub CLI to post a comment, we set the GITHUB_TOKEN environment variable to the value of the GITHUB_TOKEN secret, which is an installation access token, created when the workflow runs. We set the PR_URL environment variable to the URL of the newly created pull request, and we use this in the gh command.